pyproject.toml
[project]
name = "study-langchain-v1"
version = "0.1.0"
description = "stydy langchain 1.2.11"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.11",
    "langchain-anthropic>=1.3.4",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-mistralai>=1.1.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]
